pyproject.toml
[tool.poetry]
name = "oauth-agent-ex"
version = "0.1.0"
description = "An OAuth 2.0 extension for AI agent authentication"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
pyjwt = "^2.6.0"
cryptography = "^40.0.2"
requests = "^2.30.0"
pydantic = "^1.10.7"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
ignore_missing_imports = true
strict_optional = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
